water record code refactoring
